--- a/src/hotspot/cpu/riscv/stubRoutines_riscv.hpp
+++ b/src/hotspot/cpu/riscv/stubRoutines_riscv.hpp
@@ -36,8 +36,8 @@ static bool returns_to_call_stub(address return_pc) {
 }
 
 enum platform_dependent_constants {
-  code_size1 = 19000,          // simply increase if too small (assembler will crash if too small)
-  code_size2 = 28000           // simply increase if too small (assembler will crash if too small)
+  code_size1 = 20000,          // simply increase if too small (assembler will crash if too small)
+  code_size2 = 30000           // simply increase if too small (assembler will crash if too small)
 };
 
 class riscv {