@RealFYang 2023-09-20
PS: I just realized that this fix is not there in jdk21u. I think this should be backported to jdk21u first. Then we can consider jdk17u-dev backport.