@Ilyagavrilin 2023-09-21
> Looks good to me. Any tests performed?

Yes, some performance results (measured on thead board):
Without intrinsic: 
```
Benchmark                    (digesterName)  (length)  (provider)   Mode  Cnt    Score   Error   Units
MessageDigests.digest                   md5        64     DEFAULT  thrpt   50  731.274 ± 1.250  ops/ms
MessageDigests.digest                   md5      1024     DEFAULT  thrpt   50  107.604 ± 0.164  ops/ms
MessageDigests.digest                   md5     16384     DEFAULT  thrpt   50    7.301 ± 0.026  ops/ms
MessageDigests.getAndDigest             md5      1024     DEFAULT  thrpt   50   96.412 ± 0.517  ops/ms
MessageDigests.getAndDigest             md5     16384     DEFAULT  thrpt   50    7.251 ± 0.039  ops/ms
```
With intrinsic:
```
Benchmark                    (digesterName)  (length)  (provider)   Mode  Cnt    Score   Error   Units
MessageDigests.digest                   md5        64     DEFAULT  thrpt   50  772.265 ± 3.972  ops/ms
MessageDigests.digest                   md5      1024     DEFAULT  thrpt   50  120.391 ± 0.363  ops/ms
MessageDigests.digest                   md5     16384     DEFAULT  thrpt   50    8.294 ± 0.012  ops/ms
MessageDigests.getAndDigest             md5      1024     DEFAULT  thrpt   50  107.167 ± 0.102  ops/ms
MessageDigests.getAndDigest             md5     16384     DEFAULT  thrpt   50    8.170 ± 0.036  ops/ms
```
Also, tier1 and hotspot:tier2 passed.