@Ilyagavrilin 2023-09-21
> PS: I just realized that this fix is not there in jdk21u. I think this should be backported to jdk21u first. Then we can consider jdk17u-dev backport.

Created pr to jdk21u with MD5 backport: https://github.com/openjdk/jdk21u/pull/184 . Backport to jdk21u clean.