@Ilyagavrilin 2023-09-23
/approval request I would like to backport this into jdk17u; Applied to jdk21u, low-risk, risc-v only change, Testing: tier1/hotspot:tier2 on risc-v board